feat: внесены изменения после код-ревью
