+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -1,11 +1,11 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ReviewNotFoundException;
 import ru.yandex.practicum.filmorate.model.*;
 import ru.yandex.practicum.filmorate.storage.event.EventStorage;
-import ru.yandex.practicum.filmorate.storage.review.ReviewRatingsDao;
 import ru.yandex.practicum.filmorate.storage.review.ReviewStorage;
 
 import java.util.Collection;
@@ -13,31 +13,23 @@ import java.util.NoSuchElementException;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class ReviewService {
 
     private final ReviewStorage reviewStorage;
-    private final ReviewRatingsDao reviewRatingsDao;
     private final EventStorage eventStorage;
 
-    public ReviewService(ReviewStorage reviewStorage,
-                         ReviewRatingsDao reviewRatingsDao,
-                         EventStorage eventStorage){
-        this.reviewStorage = reviewStorage;
-        this.reviewRatingsDao = reviewRatingsDao;
-        this.eventStorage = eventStorage;
-    }
-
-    public Collection<Review> getAllReviews(Long filmId, Long count){
+    public Collection<Review> getAllReviews(Long filmId, Long count) {
         return reviewStorage.getAll(filmId, count);
     }
 
-    public Review getReviewById(final Long id){
+    public Review getReviewById(final Long id) {
         return reviewStorage.getById(id).orElseThrow(() ->
                 new ReviewNotFoundException(String.format("Attempt to get review with absent id = %d",
-                       id)));
+                        id)));
     }
 
-    public Review addReview(Review review){
+    public Review addReview(Review review) {
         Review rev = reviewStorage.add(review);
         eventStorage.addNewEvent(new Event.Builder()
                 .setCurrentTimestamp()
@@ -49,7 +41,7 @@ public class ReviewService {
         return rev;
     }
 
-    public Review updateReview(Review review){
+    public Review updateReview(Review review) {
         if (reviewStorage.isReviewExists(review.getReviewId())) {
             Review rev = reviewStorage.getById(review.getReviewId()).get();
             eventStorage.addNewEvent(new Event.Builder()
@@ -63,7 +55,7 @@ public class ReviewService {
         return reviewStorage.update(review);
     }
 
-    public void deleteReviewById(Long id){
+    public void deleteReviewById(Long id) {
         Review review = null;
         try {
             review = reviewStorage.getById(id).get();
@@ -80,19 +72,19 @@ public class ReviewService {
                 .build());
     }
 
-    public void addUserLike(Long id, Long userId){
-        reviewRatingsDao.addLike(id, userId);
+    public void addUserLike(Long id, Long userId) {
+        reviewStorage.addLike(id, userId);
     }
 
-    public void addUserDislike(Long id, Long userId){
-        reviewRatingsDao.addDislike(id, userId);
+    public void addUserDislike(Long id, Long userId) {
+        reviewStorage.addDislike(id, userId);
     }
 
-    public void deleteUserLike(Long id, Long userId){
-        reviewRatingsDao.deleteLike(id, userId);
+    public void deleteUserLike(Long id, Long userId) {
+        reviewStorage.deleteLike(id, userId);
     }
 
-    public void deleteUserDislike(Long id, Long userId){
-        reviewRatingsDao.deleteDislike(id, userId);
+    public void deleteUserDislike(Long id, Long userId) {
+        reviewStorage.deleteDislike(id, userId);
     }
 }
